Refactor one-time boot repairs
