+++ server/jobs/README.md
@@ -1,0 +1,3 @@
+# Server Jobs
+
+Scheduled and background job modules belong here as startup is split out of `server/index.ts`.
